+++ desktop/php/weenect.php
@@ -138,6 +138,18 @@ $confList = weenect::W_CONF_common;
 									<input type="checkbox" class="eqLogicAttr form-control" data-l1key="configuration" data-l2key="show_history" checked>
 								</div>
 							</div>
+							<div class="form-group">
+								<label class="col-sm-4 control-label">{{Zoom par défaut}}
+								<sup><i class="fas fa-question-circle tooltips" title="{{zoom par défaut lors de l'affichage de la tuile}}"></i></sup>
+								</label>
+								<div class="col-sm-4">
+									<input type="range" class="eqLogicAttr form-control" data-l1key="configuration" data-l2key="default_zoom" min="10" max="20" step="1">
+									
+								</div>
+								<div class="col-sm-2">
+									<span id="default_zoom_value"></span>
+								</div>
+							</div>
 
 							
 						</div>